// src/app/interfaces/users-response-interface/users-response.ts
import { AddressList } from "../../types/address-list";
import { DependentsList } from "../../types/dependents-list";
import { PhoneList } from "../../types/phone-list";

export interface IUsersResponse {
  name: string;
  email: string;
  country: string;
  state: string;
  maritalStatus: number;
  monthlyIncome: number;
  birthDate: string;
  phoneList: PhoneList;
  addressList: AddressList;
  dependentsList: DependentsList;
}